Refactor source organization
