+++ README.md
@@ -38,21 +38,24 @@ The application follows the Uncle Bob "[Clean Architecture](https://8thlight.com
 app 
  └ lib                              → Application sources 
     └ application                   → Application services layer
-       └ repositories               → Data access objects interfaces (unfortunately, there is no Interface pattern in JavaScript)
        └ security                   → Security tools interfaces (ex: AccessTokenManager.js, to generate and decode OAuth access token)
        └ use_cases                  → Application business rules 
-    └ domain                        → Enterprise core business layer
-       └ models                     → Domain model objects such as Entities, Aggregates, Value Objects, Business Events, etc.
-       └ services                   → Domain services, e.g. business objects that manipulate multiple and different Domain Models
+    └ domain                        → Enterprise core business layer such as domain model objects (Aggregates, Entities, Value Objects) and repository interfaces
     └ infrastructure                → Frameworks, drivers and tools such as Database, the Web Framework, mailing/logging/glue code etc.
-       └ database                   → ORM and database connection objects
+       └ config                     → Application configuration files, modules and services
+          └ service-locator.js      → Module that manage service implementations by environment
+       └ orm                        → Database ORMs middleware (Sequelize for SQLite3 or PostgreSQL, Mongoose for MongoDB)
+          └ mongoose                → Mongoose client and schemas
+          └ sequelize               → Sequelize client and models
+       └ repositories               → Implementation of domain repository interfaces
+       └ security                   → Security tools implementations (ex: JwtAccessTokenManager)
        └ webserver                  → Hapi.js Web server configuration (server, routes, plugins, etc.)
+          └ oauth                   → Hapi.js authentication strategies and schemas
           └ server.js               → Hapi.js server definition
     └ interfaces                    → Adapters and formatters for use cases and entities to external agency such as Database or the Web
        └ controllers                → Hapi.js route handlers
-       └ security                   → Security tools implementations (ex: JwtAccessTokenManager)
+       └ routes                     → Hapi.js route definitions
        └ serializers                → Converter objects that transform outside objects (ex: HTTP request payload) to inside objects (ex: Use Case request object)
-       └ storage                    → Repository implementations
  └ node_modules (generated)         → NPM dependencies
  └ test                             → Source folder for unit or functional tests
  └ index.js                         → Main application entry point
