+++ lib/infrastructure/config/service-locator.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const JwtAccessTokenManager = require('./security/JwtAccessTokenManager');
-const UserSerializer = require('./serializers/UserSerializer');
-const UserRepositoryInMemory = require('./storage/UserRepositoryInMemory');
-const UserRepositorySQLite= require('./storage/UserRepositorySQLite');
-const UserRepositoryMongo = require('./storage/UserRepositoryMongo');
+const JwtAccessTokenManager = require('../security/JwtAccessTokenManager');
+const UserSerializer = require('../../interfaces/serializers/UserSerializer');
+const UserRepositoryInMemory = require('../repositories/UserRepositoryInMemory');
+const UserRepositorySQLite= require('../repositories/UserRepositorySQLite');
+const UserRepositoryMongo = require('../repositories/UserRepositoryMongo');
 
 function buildBeans() {
