+++ lib/infrastructure/repositories/UserRepositoryInMemory.js
