+++ lib/infrastructure/repositories/UserRepositoryMongo.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const User = require('../../domain/User');
-const MongooseUser = require('../../infrastructure/orm/mongoose/schemas/User');
+const MongooseUser = require('../orm/mongoose/schemas/User');
 const UserRepository = require('../../domain/UserRepository');
 
 module.exports = class extends UserRepository {
