+++ lib/infrastructure/repositories/UserRepositorySQLite.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const sequelize = require('../../infrastructure/orm/sequelize/sequelize');
+const sequelize = require('../orm/sequelize/sequelize');
 const User = require('../../domain/User');
 const UserRepository = require('../../domain/UserRepository');
