+++ lib/infrastructure/security/JwtAccessTokenManager.js
