+++ lib/infrastructure/webserver/server.js
@@ -55,9 +55,9 @@ const createServer = async () => {
   // Register custom plugins
   await server.register([
     require('./oauth'),
-    require('./routes/hello'),
-    require('./routes/private'),
-    require('./routes/users'),
+    require('../../interfaces/routes/hello'),
+    require('../../interfaces/routes/private'),
+    require('../../interfaces/routes/users'),
   ]);
 
   return server;
