+++ lib/interfaces/controllers/AuthorizationController.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Boom = require('@hapi/boom');
-const serviceLocator = require('../service-locator');
+const serviceLocator = require('../../infrastructure/config/service-locator');
 const GetAccessToken = require('../../application/use_cases/GetAccessToken');
 const VerifyAccessToken = require('../../application/use_cases/VerifyAccessToken');
