+++ lib/interfaces/controllers/UsersController.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Boom = require('@hapi/boom');
-const serviceLocator = require('../service-locator');
+const serviceLocator = require('../../infrastructure/config/service-locator');
 const ListUsers = require('../../application/use_cases/ListUsers');
 const CreateUser = require('../../application/use_cases/CreateUser');
 const GetUser = require('../../application/use_cases/GetUser');
