+++ lib/interfaces/routes/hello.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const HelloController = require('../../../interfaces/controllers/HelloController');
+const HelloController = require('../controllers/HelloController');
 
 module.exports = {
   name: 'hello',
