+++ lib/interfaces/routes/users.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const UsersController = require('../../../interfaces/controllers/UsersController');
+const UsersController = require('../controllers/UsersController');
 
 module.exports = {
   name: 'users',
